small improvements / restructuring
